Refactored network message structure and updated network tests

+++ src/network/include/network/message_header.hpp
@@ -5,5 +5,5 @@
 
 struct MessageHeader {
     MessageType type;
-    std::uint64_t payload_size;
+    std::uint16_t payload_size;
 };
